Make toolchain selfcontained (#28)
* Make toolchain selfcontained

Multiarch builds need --copts and such to not be crutial, or be the same
on all platforms.  Given we need to support running on Windows, OSX and
Linux, those won't work.  Instead, push all those flags into the
crosstool.

If we want to be configurable here, we can take arguments in the
WORKSPACE file to tweak the crosstool instead.

Signed-off-by: Austin Schuh <austin.linux@gmail.com>

* Address comments and add feature for PJ

* Really move werror over

---------

Signed-off-by: Austin Schuh <austin.linux@gmail.com>